+++ src/ActionProvider.js
@@ -1,75 +1,59 @@
 import React, { PropTypes } from 'react';
-import Promise from 'Promise';
-import ActionProxy from './ActionProxy';
 
 const Identity = x => x;
-const createActionProxy = alias => new ActionProxy(alias);
 
 const ActionProvider = function(Component, {
   actions,
   displayName,
   reducer = Identity,
   serviceWrapper = Identity,
-  verbose = false,
   passDispatchProp = true,
 }) {
-  const debugLog = verbose ? console.debug.bind(console) : Function.prototype;
-
-  return React.createClass({
-    displayName: displayName || `ActionProvider(${Component.displayName})`,
-
-    contextTypes: {
-      availableActions: PropTypes.arrayOf(PropTypes.string),
-      dispatch: PropTypes.func,
-    },
-
-    childContextTypes: {
-      availableActions: PropTypes.arrayOf(PropTypes.string),
-      dispatch: PropTypes.func,
-    },
-
+  const actionNames = Object.keys(actions);
+  class WithActions extends React.Component {
     getChildContext() {
       return {
-        availableActions: (this.context.availableActions || []).concat(Object.keys(actions)),
-        dispatch: this.dispatchAction,
+        availableActions: (this.context.availableActions || []).concat(actionNames),
+        dispatch: this.bindings.dispatchAction,
       }
-    },
+    }
 
-    componentWillMount() {
+    constructor() {
+      super()
+
       this.actionBuffer = [];
-    },
+      this.bindings = {
+        dispatchAction: this.dispatchAction.bind(this)
+      };
+    }
 
     componentDidMount() {
       this.flushQueuedActions();
-    },
+    }
 
     componentDidUpdate() {
       this.flushQueuedActions();
-    },
+    }
 
     componentWillUnmount() {
       // TODO: possible leak here? what about pending promises?
       this.actionBuffer = null;
-    },
+    }
 
     render() {
       const decoratorProps = {};
 
       if (passDispatchProp) {
-        decoratorProps.dispatch = this.dispatchAction;
+        decoratorProps.dispatch = this.bindings.dispatchAction;
       }
 
       return <Component ref="container" {...decoratorProps} {...this.props} />
-    },
+    }
 
-    dispatchAction(type, payload) {
+    dispatchAction(type, ...payload) {
       // Not an action we provide? Propagate
       if (!actions.hasOwnProperty(type)) {
-        return this.propagateAction(type, payload);
-      }
-      // An alias? resolve it and re-dispatch
-      else if (actions[type] instanceof ActionProxy) {
-        return this.dispatchAction(actions[type].type, payload);
+        return this.propagateAction(type, ...payload);
       }
       // Not ready yet? Queue it until we are. This happens if a child emitter
       // dispatches an action during the componentWillMount or componentDidMount
@@ -82,22 +66,18 @@ const ActionProvider = function(Component, {
         const actionHandler = actions[type];
         const state = reducer(this.refs.container);
 
-        debugLog(`Dispatching action "${type}":`, payload);
-
         // TODO: validate payload
 
         return serviceWrapper(
-          actionHandler(state, payload, {
-            propagate: this.propagateAction.bind(null, type, payload),
-            dispatch: this.dispatchAction,
+          actionHandler(state, ...payload, {
+            propagate: this.propagateAction.bind(this, type, ...payload),
+            dispatch: this.bindings.dispatchAction,
           })
         );
       }
-    },
+    }
 
     dispatchActionWhenReady(type, payload) {
-      debugLog(`Deferring action "${type}" until container is ready...`);
-
       return new Promise((resolve, reject) => {
         if (!this.isMounted()) {
           reject();
@@ -106,28 +86,35 @@ const ActionProvider = function(Component, {
           this.actionBuffer.push({ type, payload, promise: { resolve, reject } });
         }
       });
-    },
+    }
 
-    propagateAction(type, payload) {
+    propagateAction(type, ...payload) {
       if (this.context.dispatch) {
-        return this.context.dispatch(type, payload);
+        return this.context.dispatch(type, ...payload);
       }
       else {
         return Promise.reject(new Error(`Unknown action "${type}"`));
       }
-    },
+    }
 
     flushQueuedActions() {
       this.actionBuffer.splice(0).forEach(({ type, payload, promise }) => {
-        debugLog(`Dispatching deferred action "${type}".`);
-
-        this.dispatchAction(type, payload).then(promise.resolve, promise.reject);
+        this.dispatchAction(type, ...payload).then(promise.resolve, promise.reject);
       });
     }
-  })
+  };
+
+  WithActions.displayName = displayName || `ActionProvider(${Component.displayName})`;
+  WithActions.contextTypes = {
+    availableActions: PropTypes.arrayOf(PropTypes.string),
+    dispatch: PropTypes.func,
+  };
+  WithActions.childContextTypes = {
+    availableActions: PropTypes.arrayOf(PropTypes.string),
+    dispatch: PropTypes.func,
+  };
+
+  return WithActions;
 }
 
-ActionProvider.ActionProxy = createActionProxy;
-
 export default ActionProvider;
-export { createActionProxy as ActionProxy };