+++ src/ActionProxy.js
@@ -1,3 +1,0 @@
-export default function ActionProxy(alias) {
-  this.type = alias;
-}