+++ src/index.js
@@ -1,10 +1,7 @@
-import ActionProvider, { ActionProxy } from './ActionProvider';
+import ActionProvider from './ActionProvider';
 import ActionEmitter from './ActionEmitter';
-import { PropTypes as Types } from 'react-schema';
 
 export {
   ActionProvider,
-  ActionProxy,
   ActionEmitter,
-  Types,
 };